another 5 or so

+++ Libft_xD/ft_strchr.c
@@ -6,14 +6,14 @@
 /*   By: duk <duk@student.42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/08/26 18:31:15 by duk               #+#    #+#             */
-/*   Updated: 2026/08/26 20:05:17 by duk              ###   ########.fr       */
+/*   Updated: 2026/08/27 14:38:25 by duk              ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 char	*ft_strchr(const char *s, int c)
 {
 	unsigned char	uc;
-	int				i;
+	size_t			i;
 
 	i = 0;
 	uc = (unsigned char)c;
